Fix #27: navigation puck falls off route when panning
The bug was especially noticeable when panning quickly or zooming in
while navigating.

The issue is that the upstream method signature changed in
76469a0c2227927f03a15fb95b3a20f34becc78c and then again in
8ad5f9c1fd58c60b93d26c2819637fd8f2144fd1, which first appeared in the
ios-v6.0.0 tag.

So our "overridden method" wasn't actually being called.

It's not the commit authors fault that it broke - this is a private
method we're calling.

A longer term fix might be to include some kind of public stable
delegate method rather than an override.