Show end of route even if remainingVoiceInstructions is nil
Previously, having voice instructions was a requirement of ending the
route.

Now we'll end the route when banner instructions are done *and* only
consider voice instructions if there were any to begin with.

Note: This currently crashes do to another issue with the end of route
view controller: see https://github.com/maplibre/maplibre-navigation-ios/issues/57