Restore the ability to init NavVC without animations.
Prior to #54, the NavigationViewController could only be used modally.

With #54, you can switch between "map mode" and "navigation mode". To do
so, an animation was introduced for the sake of smoothly transitioning from "map mode" to "navigation mode".

However, if you are still using the NavigationViewController modally,
this animation might interfere with other things you were doing - like
setting the camera.