Cleanup: remove private header
This header was copied in from MapLibre Native to expose a private
method.

Previously this got out of sync and caused a bug:
https://github.com/maplibre/maplibre-navigation-ios/pull/53

Instead we can use the already public delegate method.

Note: The deleted header was importing Maplibre/Mapbox.h, but we are
using that elsewhere, so I moved that import up a level.