Work around type-checking error present in XCode 16.0
This seems like a swift/xcode bug, but the work around is harmless
enough

This assignment should be a no-op, but it avoids hitting a type checking error I'm seeing in XCode Version 16.0 (16A242d) (stable release)
I did not experience this error on XCode 15.4

The error is:
> Cannot convert return expression of type 'Dictionary<Int, Double>' to return type 'Dictionary<String, Optional<JSONValue>>.RawValue' (aka 'Dictionary<String, Optional<Any>>')

JSONValue is defined in turf.
Maybe the complexity of the various expressible-by-literal's in JSONValue/JSONObject are leading to a compiler edgecase? Just a guess.